changing parameters from size_t to int in sceneManager

+++ constants.h
@@ -9,7 +9,7 @@ namespace constants
     const double kSoundSpeed = 343.216; // This is measured in 20'C at 1000 hPa
     const double kPi = std::acos(-1);
     const double kDefaultSimulationRadius = 4; // [m] // TODO: deletethis dependency
-    const size_t kPopulation = 37; // TODO: delete this dependency
+    const int kPopulation = 37; // TODO: delete this dependency
     const double kHitAccuracy = 0.0001;
 
     // ( 2 *  pi * R ) / 2 is for dividing whole sphereWall to equal distances
